+++ flutter-hms-account/example/lib/screens/auth_screen.dart
@@ -1,11 +1,11 @@
 /*
     Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
 
-    Licensed under the Apache License, Version 2.0 (the "License");
+    Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+        https://www.apache.org/licenses/LICENSE-2.0
 
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,13 +14,9 @@
     limitations under the License.
 */
 
-import 'package:huawei_account/helpers/scope.dart';
-import 'package:huawei_account/hms_account.dart';
-import 'package:huawei_account/auth/auth_huawei_id.dart';
-import 'package:huawei_account/authbutton/huawei_id_auth_button.dart';
-import 'package:huawei_account_example/ui/auth_button.dart';
+import 'package:huawei_account/huawei_account.dart';
 import 'package:flutter/material.dart';
-import 'package:huawei_account/helpers/auth_param_helper.dart';
+import '../ui/auth_button.dart';
 
 class AuthScreen extends StatefulWidget {
   @override
@@ -32,10 +28,16 @@ class _AuthScreenState extends State<AuthScreen> {
 
   List<String> logs = [];
 
+  HmsAuthHuaweiId _id;
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         key: _scaffoldKey,
+        appBar: AppBar(
+          title: Text("HMS Account Example"),
+          backgroundColor: Colors.blueAccent,
+        ),
         body: SafeArea(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.spaceAround,
@@ -43,117 +45,101 @@ class _AuthScreenState extends State<AuthScreen> {
               SizedBox(height: 20),
               Column(
                 children: [
-                  HuaweiIdAuthButton(onPressed: _signIn, buttonColor: AuthButtonBackground.RED),
-                  authButton("SIGN IN WITH AUTHORIZATION CODE", _signInWithAuthorizationCode),
+                  HmsAuthButton(
+                      onPressed: _signIn,
+                      buttonColor: AuthButtonBackground.RED),
                   authButton("SILENT SIGN IN", _silentSignIn),
                   authButton("SIGN OUT", _signOut),
                   authButton("REVOKE AUTHORIZATION", _revokeAuthorization),
-                  authButton("SMS VERIFICATION", _smsVerification),
+                  authButton("SMS VERIFICATION", _smsVerification)
                 ],
               ),
+              Divider(indent: 15, endIndent: 15, color: Colors.blueGrey),
               Expanded(
                   child: GestureDetector(
-                    onDoubleTap: () {
-                      setState(() {
-                        logs.clear();
-                      });
+                onDoubleTap: () {
+                  setState(() {
+                    logs.clear();
+                  });
+                },
+                child: Container(
+                  child: ListView.builder(
+                    itemCount: logs.length,
+                    itemBuilder: (context, index) {
+                      return Padding(
+                        padding: const EdgeInsets.all(15),
+                        child: Text(logs[index],
+                            style: TextStyle(color: Colors.black54)),
+                      );
                     },
-                    child: Container(
-                      child: ListView.builder(
-                        itemCount: logs.length,
-                        itemBuilder: (context, index) {
-                          return Padding(
-                            padding: const EdgeInsets.all(15),
-                            child: Text(logs[index], style: TextStyle(color: Colors.black54)),
-                          );
-                        },
-                      ),
-                    ),
-                  )
-              ),
+                  ),
+                ),
+              )),
             ],
           ),
         ));
   }
 
   _signIn() async {
-    // BUILD DESIRED PARAMS
-    AuthParamHelper authParamHelper = new AuthParamHelper();
-    authParamHelper..setIdToken()..setAuthorizationCode()..setAccessToken()..setProfile()..setEmail()..setId()..addToScopeList([Scope.openId])..setRequestCode(8888);
-    // GET ACCOUNT INFO FROM PLUGIN
+    // This parameter is optional. You can run the method with default options.
+    final helper = new HmsAuthParamHelper();
+    helper
+      ..setIdToken()
+      ..setAccessToken()
+      ..setAuthorizationCode()
+      ..setEmail()
+      ..setProfile();
     try {
-      final AuthHuaweiId accountInfo = await HmsAccount.signIn(authParamHelper);
-      setState(() {
-        logs.add(accountInfo.displayName);
-      });
-    } on Exception catch(exception) {
-      print(exception.toString());
+      _id = await HmsAuthService.signIn(authParamHelper: helper);
+      _addToLogs("FROM SIGN IN: ${_id.displayName}");
+    } on Exception catch (e) {
+      print(e.toString());
     }
-    /// TO VERIFY ID TOKEN, AuthParamHelper()..setIdToken()
-    //performServerVerification(accountInfo.idToken);
   }
 
   _signOut() async {
-    final signOutResult = await HmsAccount.signOut();
-    if (signOutResult) {
-      setState(() {
-        logs.add("Sign out success");
-      });
-    } else {
-      setState(() {
-        logs.add("Sign out failed");
-      });
+    try {
+      final bool result = await HmsAuthService.signOut();
+      _addToLogs("FROM SIGN OUT: $result");
+    } on Exception catch (e) {
+      print(e.toString());
     }
   }
 
   _silentSignIn() async {
-    AuthParamHelper authParamHelper = new AuthParamHelper();
     try {
-      final AuthHuaweiId accountInfo = await HmsAccount.silentSignIn(authParamHelper);
-      setState(() {
-        logs.add(accountInfo.displayName);
-      });
-    } on Exception catch(exception) {
-      print(exception.toString());
+      final HmsAuthHuaweiId id = await HmsAuthService.silentSignIn();
+      _addToLogs("FROM SILENT SIGN IN: ${id.displayName}");
+    } on Exception catch (e) {
+      print(e.toString());
     }
   }
 
   _revokeAuthorization() async {
-    final bool revokeResult = await HmsAccount.revokeAuthorization();
-    if (revokeResult) {
-      setState(() {
-        logs.add("Revoked Auth Successfuly");
-      });
-    } else {
-      setState(() {
-        logs.add("Failed to Revoked Auth");
-      });
-    }
-  }
-
-  _signInWithAuthorizationCode() async {
-    AuthParamHelper authParamHelper = new AuthParamHelper();
-    authParamHelper..setAuthorizationCode()..setRequestCode(1002);
     try {
-      final AuthHuaweiId accountInfo = await HmsAccount.signInWithAuthorizationCode(authParamHelper);
-      setState(() {
-        logs.add(accountInfo.authorizationCode);
-      });
-    } on Exception catch(exception) {
-      print(exception.toString());
+      final bool result = await HmsAuthService.revokeAuthorization();
+      _addToLogs("FROM REVOKE AUTHORIZATION: $result");
+    } on Exception catch (e) {
+      print(e.toString());
     }
   }
 
-  _smsVerification() async{
-    HmsAccount.smsVerification(({errorCode, message}){
+  _smsVerification() async {
+    HmsSmsManager.smsVerification(({errorCode, message}) {
       if (message != null) {
         setState(() {
-          logs.add(message);
+          _addToLogs(message);
         });
         print("Received message: $message");
       } else {
         print("Error: $errorCode");
       }
+    });
+  }
+
+  _addToLogs(String s) {
+    setState(() {
+      logs.add(s);
     });
   }
 }