+++ flutter-hms-account/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 5.0.3+302
+
+* Added HmsAuthService module.
+* Added HmsAuthManager module.
+* Added HmsAuthTool module.
+* Added HmsNetworkTool module.
+* Added HmsSmsManager module.
+* Added enableLogger and disableLogger methods to enable & disable the HMS Plugin Method Analytics.
+
 ## 5.0.0+300
 
 * Added repository url.